pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hms2"
version = "2.0.0"
description = "Another annotation-free whole-slide training approach to pathological classification."
license = "CC BY-NC-SA 4.0"
authors = [ "Chi-Chung Chen <chenchc@aetherai.com>",]

[tool.nitpick]
style = [ "https://github.com/aetherAI/Pypi-Project-Template/blob/v4.0.3/nitpick-style/lint.toml", "https://github.com/aetherAI/Pypi-Project-Template/blob/v4.0.3/nitpick-style/coverage.toml",]
cache = "never"

[tool.black]
line-length = 119
target-version = [ "py39",]
skip-string-normalization = true

[tool.pyright]
stubPath = ".stubs"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.21.6"
numba = "^0.60.0"
opencv-python = "^4.7.0.71"
scipy = "^1.7"
Pillow = "^9.1.0"
scikit-image = "^0.19.2"
requests = "^2.27.1"
scikit-learn = "^1.0.2"
tqdm = "^4.64.0"
mpi4py = "^4.1.1"
PyYAML = "^6.0"
pandas = "^1.3.5"
lifelines = "^0.27.0"
scikit-build = "^0.14.1"
scoping = "^0.1.2"
pydantic = "^1.10.2"
devtools = "^0.9.0"
atomicwrites = "^1.4.1"
einops = "^0.6.0"
rich = "^13.3.5"
onnxscript = "0.1.0.dev20231206"
shapely = "^2.0.5"
rasterio = "^1.3.10"
typing-extensions = "^4.15.0"
openslide-python = "^1.1.2"

[tool.poetry.extras]
rpy2 = [ "rpy2",]

[tool.poe.tasks]
install = [ "install-torch", "install-torchvision", "install-horovod", "install-post",]
install-torch = "python -m pip install torch==2.1.2+cu121 -f https://download.pytorch.org/whl/torch"
install-torchvision = "python -m pip install torchvision==0.16.2+cu121 -f https://download.pytorch.org/whl/torchvision"
install-horovod = "python -m pip install git+https://github.com/horovod/horovod.git@3a31d933a13c7c885b8a673f4172b17914ad334d --no-cache-dir"
install-post = "python scripts/post_install.py"
install-peft = "python -m pip install peft==0.6.0"

[tool.pytest.ini_options]
addopts = [ "--ignore=hms2/pipeline/test_emb.py",]

[tool.poetry.dependencies.rpy2]
version = "^3.5.3"
optional = true

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pytest = "^7.1.1"
isort = "^5.10.1"
mypy = "^0.942"
poethepoet = "^0.13.1"
types-PyYAML = "^6.0.5"
onnxruntime-gpu = "^1.10.0"
cython = "^3.0.0"
toml = "^0.10.2"
wemake-python-styleguide = ">=0.17"
pytest-cov = "*"

